+++ SSMP/Animation/AnimationEffect.cs
@@ -1,5 +1,6 @@
 using SSMP.Game.Settings;
 using SSMP.Internals;
+using SSMP.Util;
 using UnityEngine;
 
 namespace SSMP.Animation;
@@ -25,7 +26,7 @@ internal abstract class AnimationEffect : IAnimationEffect {
     }
 
     /// <summary>
-    /// Locate the damages_enemy FSM and change the attack direction to the given direciton. This will ensure that
+    /// Locate the damages_enemy FSM and change the attack direction to the given direction. This will ensure that
     /// enemies are getting knocked back in the correct direction from remote player's attacks.
     /// </summary>
     /// <param name="targetObject">The target GameObject to change.</param>
@@ -49,5 +50,54 @@ internal abstract class AnimationEffect : IAnimationEffect {
         // "hide" the player (assign a very small texture)
         playerObject.GetComponent<tk2dSpriteAnimator>().Stop();
         playerObject.GetComponent<tk2dSprite>().SetSprite("wall_puff0004");
+    }
+
+    /// <summary>
+    /// Gets the Effects object for a given player
+    /// </summary>
+    /// <param name="playerObject">The player using the effect</param>
+    /// <returns>The player's effects object</returns>
+    protected static GameObject GetPlayerEffects(GameObject playerObject) {
+        var effects = playerObject.FindGameObjectInChildren("Effects");
+        if (effects == null) {
+            effects = new GameObject("Effects");
+            effects.transform.SetParentReset(playerObject.transform);
+        }
+
+        return effects;
+    }
+
+    /// <summary>
+    /// Attempts to get or create an effect from the Effects sub-object
+    /// </summary>
+    /// <param name="playerObject">The player using the effect</param>
+    /// <param name="effectName">The name of the effect object</param>
+    /// <param name="effect">The effect, if found or created</param>
+    /// <returns>true if created, false otherwise</returns>
+    protected static bool TryGetEffect(GameObject playerObject, string effectName, out GameObject? effect) {
+        // Find or create effects for player
+        var effects = GetPlayerEffects(playerObject);
+
+        // Find existing effect
+        effect = effects.FindGameObjectInChildren(effectName);
+        if (effect != null) {
+            return false;
+        }
+
+        // Create new effect
+        var localEffects = HeroController.instance.gameObject.FindGameObjectInChildren("Effects");
+        if (localEffects == null) {
+            return false;
+        }
+
+        var localEffect = localEffects.FindGameObjectInChildren(effectName);
+        if (localEffect == null) {
+            return false;
+        }
+
+        effect = Object.Instantiate(localEffect, effects.transform);
+        effect.name = effectName;
+
+        return true;
     }
 }
