+++ SSMP/Animation/Effects/Tools/BaseTool.cs
@@ -1,0 +1,9 @@
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SSMP.Animation.Effects.Tools;
+
+internal abstract class BaseTool : DamageAnimationEffect {
+
+}
